Without delegation
