+++ programs/perpetuals/src/lib.rs
@@ -4,7 +4,6 @@
 
 pub mod adapters;
 pub mod error;
-pub mod governance;
 pub mod instructions;
 pub mod math;
 pub mod state;
